requirements.txt
django==3.2
djangorestframework==3.12.4
pika==1.2.0
psycopg2==2.8.6
django-cors-headers==3.7.0